requirements.txt
cycler==0.10.0
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.9.2
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
